refactor: Refactor code while reading LEM
- Updated import specifications across numerous files for clarity:, these include changes to `backend.rs`, `circuit.rs`, `path.rs`, `eval.rs`, `macros.rs`, and `pointers.rs`.
- Modified and refactored code for improved error handling, examples include changes to `interpreter.rs`, `mod.rs`, `var_map.rs`, `writer.rs`, `store.rs`.
- Included implementation of Debug traits in `var_map.rs` and `zstore.rs`.
- Enhanced handling of invalid u16 values in `Tag` implementation
- Implemented several lint checks in `.cargo/config` to ensure code quality.
- Refined match syntax in the `num_paths_taken` method on `path.rs` and `fetch_symbol_path`, `read_maybe_meta`, `hash_ptr` on `store.rs` for better performance and error handling.

+++ src/syntax_macros.rs
@@ -1,10 +1,10 @@
 #[macro_export]
 macro_rules! num {
     ($f:ty, $i:literal) => {
-        $crate::syntax::Syntax::<$f>::Num(Pos::No, ($i as u64).into())
+        $crate::syntax::Syntax::<$f>::Num(Pos::No, ($i).into())
     };
     ($i:literal) => {
-        $crate::syntax::Syntax::Num(Pos::No, ($i as u64).into())
+        $crate::syntax::Syntax::Num(Pos::No, ($i).into())
     };
     ($i:expr) => {
         $crate::syntax::Syntax::Num(Pos::No, $i)
@@ -14,10 +14,10 @@ macro_rules! num {
 #[macro_export]
 macro_rules! uint {
     ($f:ty, $i:literal) => {
-        $crate::syntax::Syntax::<$f>::UInt(Pos::No, $crate::uint::UInt::U64($i as u64))
+        $crate::syntax::Syntax::<$f>::UInt(Pos::No, $crate::uint::UInt::U64($i))
     };
     ($i:literal) => {
-        $crate::syntax::Syntax::UInt(Pos::No, $crate::uint::UInt::U64($i as u64))
+        $crate::syntax::Syntax::UInt(Pos::No, $crate::uint::UInt::U64($i))
     };
 }
